test(web): import API normalizer directly
