+++ web/src/lib/api.ts
@@ -1,17 +1,8 @@
+import { normalizeApiUrl } from "./api-url.mjs";
+
 const DEFAULT_API_BASE_URL = "https://uhsocial.in/api";
 
-export function normalizeApiUrl(baseUrl: string, path: string) {
-  const normalizedBaseUrl = baseUrl.trim().replace(/\/+$/, "");
-  const apiBaseUrl = normalizedBaseUrl.endsWith("/api")
-    ? normalizedBaseUrl
-    : `${normalizedBaseUrl}/api`;
-  const normalizedPath = path
-    .trim()
-    .replace(/^\/+/, "")
-    .replace(/^api(?:\/+|$)/, "");
-
-  return normalizedPath ? `${apiBaseUrl}/${normalizedPath}` : apiBaseUrl;
-}
+export { normalizeApiUrl };
 
 export function getApiUrl(path: string) {
   return normalizeApiUrl(
